moved stylesheets into their component folder and moved styles folder into /src
