+++ package.json
@@ -7,7 +7,8 @@
     "dev": "vite",
     "build": "vite build",
     "lint": "eslint .",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "compile:sass": "sass --watch scss:assets/css"
   },
   "dependencies": {
     "react": "^19.1.1",
