+++ src/Main.module.scss
@@ -1,14 +1,14 @@
-@use './variables.module.scss';
+@use './styles/variables.module.scss';
 
 /* ===== Elements ===== */
 :root {
-  font-family: $font-stack;
+  font-family: variables.$font-stack;
   line-height: 1.5;
   font-weight: 400;
 
   color-scheme: light dark;
   color: rgba(255, 255, 255, 0.87);
-  background-color: $body-background-color;
+  background-color: variables.$body-background-color;
 
   font-synthesis: none;
   text-rendering: optimizeLegibility;
@@ -18,12 +18,12 @@
 
 a {
   font-weight: 500;
-  color: $hyperlink-color;
+  color: variables.$hyperlink-color;
   text-decoration: inherit;
 }
 
 a:hover {
-  color: $hyperlink-color-hover;
+  color: variables.$hyperlink-color-hover;
 }
 
 body {
@@ -33,7 +33,7 @@ body {
 }
 
 h1 {
-  font-size: $font-size-h1;
+  font-size: variables.$font-size-h1;
   line-height: 1.1;
 }
 
@@ -50,7 +50,7 @@ button {
 }
 
 button:hover {
-  border-color: $hyperlink-color;
+  border-color: variables.$hyperlink-color;
 }
 
 button:focus,
