+++ src/components/ItemDetail/ItemDetail.jsx
@@ -2,7 +2,7 @@ import { useParams, Link } from 'react-router';
 import { useState, useEffect } from 'react';
 import * as itemService from '../../services/itemService';
 import { useNavigate } from 'react-router';
-import styles from "../../../public/styles/ItemDetail.module.scss";
+import styles from "./ItemDetail.module.scss";
 
 const ItemDetail = () => {
   const [item, setItem] = useState(null);
