+++ src/components/ItemDetail/ItemDetail.module.scss
@@ -1,4 +1,4 @@
-@use './variables.module.scss';
+@use '../../styles/variables.module.scss';
 /* ===== Elements ===== */
 /* ===== Classes ===== */
 /* ===== Ids ===== */
