+++ src/components/QwishForm/QwishForm.module.scss
@@ -1,4 +1,5 @@
-@use './variables.module.scss';
+@use '../../styles/variables.module.scss';
+
 /* ===== Elements ===== */
 /* ===== Classes ===== */
 /* ===== Ids ===== */
